# Span.End should return ASAP if no exporters + if not sampled