Fixes #681

* Span.End should return ASAP if no exporters are provided to
avoid the overhead of preparing data that won't be used
* If the span is not sampled, no need to
invoke makeSpanData(). The current user is the spanStore
that marks it as finished but not sure that it is used.
* Since (*Span).spanContext is a non-pointer value, it is always
zeroed as is the case when it is a root span. Only set
(*Span).spanContext = parent iff hasParent is true thus
bringing the number of sets in the case when it is a root span
always to 1 instead of 2 before.
* Update the benchmarks to include the case with zero and then one
exporter.
* Also unroll the benchmarking code as it was giving spurious/wrong
results when testing no vs with exporters despite us invoking
b.ResetTimer(). Perhaps this is just a flake, but anyways the
benchmarks aren't radically changing so for now it is okay
to have them as they are, unrolled and unrefactored, to ensure we
capture any sensitivity.

This change shows an improvement in the case where there'll be
no exporters nor data. If we want to instrument many libraries,
convincing their owners to use OpenCensus will mean that we have
to show little overhead keeping span.End() in there even when
there are no exporters attached.

Results on my non-CPU scaled Macbook Pro.
```shell
$ benchstat before.txt after.txt
name                                                old time/op    new time/op    delta
StartEndSpan_noExporters_neverSample-4                 201ns ± 3%     197ns ± 0%   -1.99%          (p=0.016 n=5+4)
StartEndSpan_noExporters_alwaysSample-4                602ns ± 1%     307ns ± 1%  -49.00%          (p=0.008 n=5+5)
StartEndSpan_withExporters_neverSample-4               200ns ± 1%     198ns ± 1%     ~             (p=0.087 n=5+5)
StartEndSpan_withExporters_alwaysSample-4              658ns ± 1%     669ns ± 0%   +1.64%          (p=0.016 n=5+5)
SpanWithAnnotations_3_noExporters_neverSample-4        261ns ± 1%     260ns ± 2%     ~             (p=0.159 n=5+5)
SpanWithAnnotations_3_noExporters_alwaysSample-4      1.30µs ± 1%    0.66µs ± 1%  -49.44%          (p=0.008 n=5+5)
SpanWithAnnotations_3_withExporters_neverSample-4      263ns ± 0%     267ns ± 8%     ~             (p=0.714 n=4+5)
SpanWithAnnotations_3_withExporters_alwaysSample-4    1.35µs ± 2%    1.36µs ± 1%     ~             (p=0.151 n=5+5)
SpanWithAnnotations_6_noExporters_neverSample-4        335ns ± 1%     331ns ± 3%     ~             (p=0.159 n=4+5)
SpanWithAnnotations_6_noExporters_alwaysSample-4      1.59µs ± 0%    0.80µs ± 1%  -49.48%          (p=0.008 n=5+5)
SpanWithAnnotations_6_withExporters_neverSample-4      332ns ± 1%     329ns ± 1%   -0.90%          (p=0.040 n=5+5)
SpanWithAnnotations_6_withExporters_alwaysSample-4    1.65µs ± 4%    1.67µs ± 1%     ~             (p=0.190 n=5+4)
SpanID_DotString_noExporters_neverSample-4             229ns ± 4%     229ns ± 2%     ~             (p=0.508 n=5+5)
SpanID_DotString_withExporters_neverSample-4           226ns ± 0%     225ns ± 2%     ~             (p=0.540 n=4+5)
SpanID_DotString_noExporters_alwaysSample-4            231ns ± 1%     226ns ± 1%   -2.23%          (p=0.016 n=4+5)
SpanID_DotString_withExporters_alwaysSample-4          231ns ± 2%     227ns ± 1%     ~             (p=0.151 n=5+5)

name                                                old alloc/op   new alloc/op   delta
StartEndSpan_noExporters_neverSample-4                  128B ± 0%      128B ± 0%     ~     (all samples are equal)
StartEndSpan_noExporters_alwaysSample-4                 576B ± 0%      352B ± 0%  -38.89%          (p=0.008 n=5+5)
StartEndSpan_withExporters_neverSample-4                128B ± 0%      128B ± 0%     ~     (all samples are equal)
StartEndSpan_withExporters_alwaysSample-4               576B ± 0%      576B ± 0%     ~     (all samples are equal)
SpanWithAnnotations_3_noExporters_neverSample-4         152B ± 0%      152B ± 0%     ~     (all samples are equal)
SpanWithAnnotations_3_noExporters_alwaysSample-4      1.27kB ± 0%    0.71kB ± 0%  -44.03%          (p=0.008 n=5+5)
SpanWithAnnotations_3_withExporters_neverSample-4       152B ± 0%      152B ± 0%     ~     (all samples are equal)
SpanWithAnnotations_3_withExporters_alwaysSample-4    1.27kB ± 0%    1.27kB ± 0%     ~     (all samples are equal)
SpanWithAnnotations_6_noExporters_neverSample-4         176B ± 0%      176B ± 0%     ~     (all samples are equal)
SpanWithAnnotations_6_noExporters_alwaysSample-4      1.30kB ± 0%    0.74kB ± 0%  -43.21%          (p=0.008 n=5+5)
SpanWithAnnotations_6_withExporters_neverSample-4       176B ± 0%      176B ± 0%     ~     (all samples are equal)
SpanWithAnnotations_6_withExporters_alwaysSample-4    1.30kB ± 0%    1.30kB ± 0%     ~     (all samples are equal)
SpanID_DotString_noExporters_neverSample-4             56.0B ± 0%     56.0B ± 0%     ~     (all samples are equal)
SpanID_DotString_withExporters_neverSample-4           56.0B ± 0%     56.0B ± 0%     ~     (all samples are equal)
SpanID_DotString_noExporters_alwaysSample-4            56.0B ± 0%     56.0B ± 0%     ~     (all samples are equal)
SpanID_DotString_withExporters_alwaysSample-4          56.0B ± 0%     56.0B ± 0%     ~     (all samples are equal)

name                                                old allocs/op  new allocs/op  delta
StartEndSpan_noExporters_neverSample-4                  2.00 ± 0%      2.00 ± 0%     ~     (all samples are equal)
StartEndSpan_noExporters_alwaysSample-4                 4.00 ± 0%      3.00 ± 0%  -25.00%          (p=0.008 n=5+5)
StartEndSpan_withExporters_neverSample-4                2.00 ± 0%      2.00 ± 0%     ~     (all samples are equal)
StartEndSpan_withExporters_alwaysSample-4               4.00 ± 0%      4.00 ± 0%     ~     (all samples are equal)
SpanWithAnnotations_3_noExporters_neverSample-4         4.00 ± 0%      4.00 ± 0%     ~     (all samples are equal)
SpanWithAnnotations_3_noExporters_alwaysSample-4        10.0 ± 0%       7.0 ± 0%  -30.00%          (p=0.008 n=5+5)
SpanWithAnnotations_3_withExporters_neverSample-4       4.00 ± 0%      4.00 ± 0%     ~     (all samples are equal)
SpanWithAnnotations_3_withExporters_alwaysSample-4      10.0 ± 0%      10.0 ± 0%     ~     (all samples are equal)
SpanWithAnnotations_6_noExporters_neverSample-4         6.00 ± 0%      6.00 ± 0%     ~     (all samples are equal)
SpanWithAnnotations_6_noExporters_alwaysSample-4        12.0 ± 0%       9.0 ± 0%  -25.00%          (p=0.008 n=5+5)
SpanWithAnnotations_6_withExporters_neverSample-4       6.00 ± 0%      6.00 ± 0%     ~     (all samples are equal)
SpanWithAnnotations_6_withExporters_alwaysSample-4      12.0 ± 0%      12.0 ± 0%     ~     (all samples are equal)
SpanID_DotString_noExporters_neverSample-4              3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
SpanID_DotString_withExporters_neverSample-4            3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
SpanID_DotString_noExporters_alwaysSample-4             3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
SpanID_DotString_withExporters_alwaysSample-4           3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
```